refactor : domain, view package 생성 및 클래스 이동
