+++ src/main/java/racingCar/CarRacing.java
@@ -1,6 +1,8 @@
 package racingCar;
 
-import java.util.ArrayList;
+import racingCar.domain.Cars;
+import racingCar.view.InputView;
+import racingCar.view.ResultView;
 
 public class CarRacing {
     public static void main(String[] args) {
