+++ src/test/java/racingCar/CarTest.java
@@ -3,6 +3,7 @@ package racingCar;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingCar.domain.Car;
 
 import java.util.Random;
